requirements.txt
google-genai>=1.0.0
playwright>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0
beautifulsoup4>=4.12.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
